+++ network-conduit-tls/ChangeLog.md
@@ -1,3 +1,15 @@
+## 1.2.0.1
+
+* tls 1.3 support
+
+## 1.2.0
+
+* Drop system-filepath
+
+## 1.1.2
+
+* Added 'runGeneralTCPServerTLS' function [#208](https://github.com/snoyberg/conduit/pull/208)
+
 ## 1.1.1.1
 
 * Fill in `appRawSocket` for streaming-commons 0.1.12 and later